[screenlocker] Add emergency mode for greeter crashing
If the greeter crashes on startup ksld is caught in a restart loop which
causes 100 % CPU usage and makes it impossible to unlock the session.

With this change the greeter is only allowed to crash four times and
ksld goes into an emergency mode. This mode ensures that our black
background window gets shown with a message indicating how to unlock
using loginctl. This is rather technical but better than leaving the
system in a state which makes it impossible to unlock.

Directly unlocking in the error case is not an option as that could
be used to attack the screen locker infrastructure.

BUG: 351696
REVIEW: 124915

+++ ksmserver/screenlocker/ksldapp.cpp
@@ -206,7 +206,12 @@ void KSldApp::initialize()
                 return;
             }
             // failure, restart lock process
-            startLockProcess(EstablishLock::Immediate);
+            m_greeterCrashedCounter++;
+            if (m_greeterCrashedCounter < 4) {
+                startLockProcess(EstablishLock::Immediate);
+            } else if (m_lockWindow) {
+                m_lockWindow->emergencyShow();
+            }
         }
     );
     connect(m_lockProcess, static_cast<void (QProcess::*)(QProcess::ProcessError)>(&QProcess::error), this,
@@ -234,8 +239,12 @@ void KSldApp::initialize()
     connect(m_logind, &LogindIntegration::requestUnlock, this,
         [this]() {
             if (lockState() == Locked) {
-                s_logindExit = true;
-                m_lockProcess->kill();
+                if (m_lockProcess->state() != QProcess::NotRunning) {
+                    s_logindExit = true;
+                    m_lockProcess->kill();
+                } else {
+                    doUnlock();
+                }
             }
         }
     );
@@ -486,6 +495,7 @@ void KSldApp::doUnlock()
     m_lockWindow = NULL;
     m_lockState = Unlocked;
     m_lockedTimer.invalidate();
+    m_greeterCrashedCounter = 0;
     endGraceTime();
     KDisplayManager().setLock(false);
     m_waylandServer->stop();
